applycfg: Move to util package
This change moves applycfg into the util package tree, since it no longer
depends on a database connection.
